Restructure to prepare implementation of deterministic STM.
